ASGI HTTP is more-or-less done, I haven't added WebSocket support yet though
